backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
sqlmodel==0.0.18
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
alembic==1.13.1
pydantic==2.7.2
pydantic-settings==2.3.1